requirements.txt
# Core Web Framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Resume Parsing
pdfplumber>=0.11.0
python-docx==1.1.2

# NLP
spacy>=3.8.0

# Data Validation
pydantic>=2.7.0
pydantic-settings>=2.0.0

# Utilities
python-dateutil>=2.9.0
typing-extensions>=4.11.0
werkzeug>=3.0.0
